# core/API-swagger.yaml
openapi: 3.1.0
info:
  title: SupremeAI 2.0
  version: 0.1.0
paths:
...
  /api/browser/surf/navigate:
    post:
      tags:
      - browser
      summary: Navigate
      operationId: navigate_api_browser_surf_navigate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/browser/surf/click:
    post:
      tags:
      - browser
      summary: Click
      operationId: click_api_browser_surf_click_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/browser/surf/fill:
    post:
      tags:
      - browser
      summary: Fill
      operationId: fill_api_browser_surf_fill_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FillRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/browser/surf/click-at:
    post:
      tags:
      - browser
      summary: Click At
      operationId: click_at_api_browser_surf_click_at_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickAtRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/browser/surf/type-key:
    post:
      tags:
      - browser
      summary: Type Key
      operationId: type_key_api_browser_surf_type_key_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/browser/simulate-activity:
    post:
      tags:
      - browser
      summary: Simulate Activity
      operationId: simulate_activity_api_browser_simulate_activity_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties:
                type: string
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/byoc/credentials:
    post:
      tags:
      - BYOC Management
      summary: Save Credentials
      description: Encrypts and saves client-provided cloud service credentials securely.
      operationId: save_credentials_api_byoc_credentials_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BYOCCredentialsPayload'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/byoc/deploy:
    post:
      tags:
      - BYOC Management
      summary: Deploy Container
      description: Checks user tier quota limits and starts background container deployment.
      operationId: deploy_container_api_byoc_deploy_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BYOCDeployRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/byoc/status/{job_id}:
    get:
      tags:
      - BYOC Management
      summary: Get Deployment Status
      description: Returns live execution state and logs for a container deployment
        job.
      operationId: get_deployment_status_api_byoc_status__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    BYOCCredentialsPayload:
      properties:
        provider:
          type: string
          enum:
          - gcp
          - aws
          - azure
          title: Provider
          default: gcp
        gcp_credentials:
          $ref: '#/components/schemas/GCPServiceAccountPayload'
      type: object
      required:
      - gcp_credentials
      title: BYOCCredentialsPayload
    BYOCDeployRequest:
      properties:
        skill_name:
          type: string
          title: Skill Name
          description: Name of the skill container to spin up
        provider:
          type: string
          enum:
          - gcp
          - aws
          - azure
          title: Provider
          default: gcp
        memory_limit:
          type: string
          title: Memory Limit
          default: 256Mi
        cpu_limit:
          type: string
          title: Cpu Limit
          default: 1000m
      type: object
      required:
      - skill_name
      title: BYOCDeployRequest
...
    ClickAtRequest:
      properties:
        x:
          type: integer
          title: X
        y:
          type: integer
          title: Y
      type: object
      required:
      - x
      - y
      title: ClickAtRequest
    ClickRequest:
      properties:
        selector:
          type: string
          title: Selector
      type: object
      required:
      - selector
      title: ClickRequest
...
    FillRequest:
      properties:
        selector:
          type: string
          title: Selector
        value:
          type: string
          title: Value
      type: object
      required:
      - selector
      - value
      title: FillRequest
...
    GCPServiceAccountPayload:
      properties:
        type:
          type: string
          title: Type
          description: Credentials Type (must be service_account)
        project_id:
          type: string
          title: Project Id
          description: GCP Project ID
        private_key_id:
          type: string
          title: Private Key Id
          description: GCP Private Key ID
        private_key:
          type: string
          title: Private Key
          description: GCP Private Key
        client_email:
          type: string
          title: Client Email
          description: GCP Client Email
      type: object
      required:
      - type
      - project_id
      - private_key_id
      - private_key
      - client_email
      title: GCPServiceAccountPayload
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KeyRequest:
      properties:
        key:
          type: string
          title: Key
      type: object
      required:
      - key
      title: KeyRequest
...
    NavigateRequest:
      properties:
        url:
          type: string
          title: Url
      type: object
      required:
      - url
      title: NavigateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login